Azure Cognitive Search: Utilizando AI Search para indexação e consulta de Dados/requirements.txt
# Azure SDK packages
azure-search-documents==11.4.0
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-ai-formrecognizer==3.3.0
azure-cognitiveservices-language-textanalytics==5.2.0

# Data processing
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0

# Web and API
requests==2.31.0
flask==3.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Jupyter and visualization
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Utilities
tqdm==4.66.1
python-dateutil==2.8.2